+++ modules/tinymce/src/themes/silver/main/ts/ui/toolbar/button/ToolbarButtons.ts
@@ -64,7 +64,8 @@ const getButtonApi = (component: AlloyComponent): Toolbar.ToolbarButtonInstanceA
   }),
   setIcon: (icon: string) => AlloyTriggers.emitWith(component, updateMenuIcon, {
     icon
-  })
+  }),
+  setTooltip: (tooltip: string) => Attribute.setAll(component.element, { 'title': tooltip, 'aria-label': tooltip })
 });
 
 const getToggleApi = (component: AlloyComponent): Toolbar.ToolbarToggleButtonInstanceApi => ({
@@ -79,7 +80,8 @@ const getToggleApi = (component: AlloyComponent): Toolbar.ToolbarToggleButtonIns
   }),
   setIcon: (icon: string) => AlloyTriggers.emitWith(component, updateMenuIcon, {
     icon
-  })
+  }),
+  setTooltip: (tooltip: string) => Attribute.setAll(component.element, { 'title': tooltip, 'aria-label': tooltip })
 });
 
 const getTooltipAttributes = (tooltip: Optional<string>, providersBackstage: UiFactoryBackstageProviders) => tooltip.map<{}>((tooltip) => ({