+++ modules/bridge/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Added `setTooltip` function to `ToolbarSplitButtonInstanceApi`. #TINY-9796
+
 ## 4.5.0 - 2023-07-12
 
 ### Added